moved utilities to manifests.utils.py
